+++ design-system/flag/examples/constellation/flags-provider.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { FlagsProvider } from '../../src';
+
+export default () => (
+  <FlagsProvider>
+    <h3>I'm wrapped in a flags provider.</h3>
+  </FlagsProvider>
+);