+++ design-system/flag/extract-react-types/show-flag-args.tsx
@@ -1,0 +1,5 @@
+import { CreateFlagArgs } from '../src';
+
+export default function ShowFlagArgs(props: CreateFlagArgs) {
+  return null;
+}