+++ editor/editor-core/src/__tests__/integration/card/inline-create/inline-create-cmd-k-keyboard.ts
@@ -23,7 +23,8 @@ BrowserTestCase(
     await page.click(`[aria-label="${messages.link.defaultMessage}"]`);
     await page.waitForSelector(linkUrlSelector);
 
-    await page.type(linkUrlSelector, ['home opt-in', 'ArrowDown', 'Return']);
+    await page.type(linkUrlSelector, 'home opt-in');
+    await page.keys(['ArrowDown', 'Return']);
     await page.waitForSelector('.inlineCardView-content-wrap');
 
     const doc = await page.$eval(editable, getDocFromElement);