+++ editor/editor-core/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @atlaskit/editor-core
 
+## 131.0.6
+
+### Patch Changes
+
+- [`91854cd5b6`](https://bitbucket.org/atlassian/atlassian-frontend/commits/91854cd5b6) - ED-10645 Fix not being able to change dates in some collab sessions
+
+## 131.0.5
+
+### Patch Changes
+
+- [`9a055964a3`](https://bitbucket.org/atlassian/atlassian-frontend/commits/9a055964a3) - ED-10628 Re-enable single click selection on mobile
+
+  Long press selection is temporarily blocked, so we will re-enable single click selection for now
+
 ## 131.0.4
 
 ### Patch Changes